requirements.txt
uvicorn
fastapi
motor
python-jose
python-multipart
openai
tiktoken